fix: separate pagination states
